Task #8697: A little oops from Tammo Jan, restoring all branches

+++ LCS/MessageBus/src/Util.cc
@@ -1,5 +1,4 @@
 #include "lofar_config.h"
-#include <Common/LofarTypes.h>
 
 #include "Util.h"
 
@@ -11,7 +10,7 @@ namespace LOFAR {
   Duration TimeOutDuration(double secs)
   {
     if (secs > 0.0)
-      return (Duration)(static_cast<uint64>(1000.0 * secs));
+      return (Duration)(1000.0 * secs);
 
     return Duration::FOREVER;
   }
